requirement.txt
python-barcode
python-barcode[images]
pyzbar